requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.8
djangorestframework==3.5.4
idna==2.6
requests==2.18.4
urllib3==1.22
wheel==0.24.0
whitenoise==3.3.1